Moved layer-compisition and light-composition-data to composition folder
