+++ src/index.js
@@ -86,7 +86,7 @@ export { DepthMaterial } from './scene/materials/depth-material.js';
 export { ForwardRenderer } from './scene/renderer/forward-renderer.js';
 export { GraphNode } from './scene/graph-node.js';
 export { Layer } from './scene/layer.js';
-export { LayerComposition } from './scene/layer-composition.js';
+export { LayerComposition } from './scene/composition/layer-composition.js';
 export { Light } from './scene/light.js';
 export { Lightmapper } from './scene/lightmapper.js';
 export { Material } from './scene/materials/material.js';
