+++ src/scene/composition/light-composition-data.js
