+++ src/scene/renderer/forward-renderer.js
@@ -34,7 +34,7 @@ import {
 import { Material } from '../materials/material.js';
 import { Mesh } from '../mesh.js';
 import { MeshInstance } from '../mesh-instance.js';
-import { LayerComposition } from '../layer-composition.js';
+import { LayerComposition } from '../composition/layer-composition.js';
 import { ShadowRenderer } from './shadow-renderer.js';
 import { Camera } from '../camera.js';
 import { GraphNode } from '../graph-node.js';
